chore(Tests): merge GeocodingService test classes together
